refactor redux: wire up termsOfServiceModalSlice
