+++ static/js/redux/constants/actionTypes.jsx
@@ -63,12 +63,6 @@ export const REMOVE_METRIC = "REMOVE_METRIC";
 export const SWITCH_METRIC = "SWITCH_METRIC";
 export const TOGGLE_METRIC_ORDER = "TOGGLE_METRIC_ORDER";
 
-// SAVE CALENDAR MODAL
-
 // UI
 export const HOVER_SEARCH_RESULT = "HOVER_SEARCH_RESULT";
 export const SET_HIGHLIGHT_NOTIFS = "SET_HIGHLIGHT_NOTIFS";
-
-// TOS
-export const TRIGGER_TOS_MODAL = "TRIGGER_TOS_MODAL";
-export const CLOSE_TOS_MODAL = "CLOSE_TOS_MODAL";
